% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component.R
\docType{class}
\name{component}
\alias{component}
\alias{Component}
\title{Class providing an object to manipulate a component of a schedule in a FR Y-9c}
\format{\code{\link{R6Class}} object.}
\usage{
component

Component(num, name, key)
}
\arguments{
\item{num}{The line item number of a component of a schedule}

\item{name}{The name of the line item number}

\item{key}{The lookup key associated with the line item number}
}
\value{
Object of \code{\link{R6Class}}

an object of class \code{component}
}
\description{
Class providing an object to manipulate a component of a schedule in a FR Y-9c
}
\section{Fields}{

\describe{
\item{\code{num}}{The line item number of a component of a schedule}

\item{\code{name}}{The name of the line item number}

\item{\code{key}}{The lookup key associated with the line item number}
}}

\section{Methods}{

\describe{
  \item{Documentation}{}
  \item{\code{new(num, name, key)}}{}
  \item{\code{initializeData(dat)}}{}
  \item{\code{add(comp)}}{}
  \item{\code{export_csv()}}{}
  \item{\code{print()}}{}
  \item{\code{getValueFromKey(key)}}{}
  \item{\code{getValueFromNum(num)}}{}
  \item{\code{getCommonSizeValueFromNum(num)}}{}
  \item{\code{getValue()}}{}
  \item{\code{getCommonSizeValue}}{}
  \item{\code{getKey()}}{}
  \item{\code{getNum()}}{}
  \item{\code{getName()}}{}
  \item{\code{getAllValues()}}{}
  \item{\code{getAllNums()}}{}
  \item{\code{getAllNames()}}{}
  \item{\code{commonSize(divisor)}}{}
}
}

\examples{
x <- component$new("1.a.", "Income", "ZZZZ1234")
Component("1.a.", "Income", "ZZZZ1234")
}
\keyword{datasets}
