% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fry9c.R
\docType{class}
\name{fry9c}
\alias{fry9c}
\alias{Fry9c}
\title{Class providing an object to manipulate a FR Y-9c and component schedules}
\format{\code{\link{R6Class}} object.}
\usage{
fry9c

Fry9c(date, omb_number, title)
}
\arguments{
\item{date}{The date associated with the FR Y-9c}

\item{omb_number}{The OMB number of the FR Y-9c documentation.  Something like "7100-0128"}

\item{title}{The title of the FR Y-9c.  Something like "Consolidated Financial Statements for Holding Companies--FR Y-9c"}
}
\value{
Object of \code{\link{R6Class}}

an object of class \code{fry9c}
}
\description{
Class providing an object to manipulate a FR Y-9c and component schedules
}
\section{Fields}{

\describe{
\item{\code{date}}{The date associated with the FR Y-9c}

\item{\code{omb_number}}{The OMB number of the FR Y-9c documentation.  Something like "7100-0128"}

\item{\code{title}}{The title of the FR Y-9c.  Something like "Consolidated Financial Statements for Holding Companies--FR Y-9c"}
}}

\section{Methods}{

\describe{
  \item{Documentation}{}
  \item{\code{new(date, omb_number, title)}}{}
  \item{\code{add(sched)}}{}
  \item{\code{getSchedule(desig)}}{}
  \item{\code{initializeData(dat)}}{}
  \item{\code{print()}}{}
  \item{\code{addBankNames(bank_names)}}{}
  \item{\code{exportExcel(output_file_name)}}{}
}
}

\examples{
x <- fry9c$new("20180331", "7100-0128",
               "Consolidated Financial Statements for Holding Companies--FR Y-9c")
Fry9c("20180331", "7100-0128",
      "Consolidated Financial Statements for Holding Companies--FR Y-9c")
}
\keyword{datasets}
