% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bank_ids.R
\name{get_bank_ids}
\alias{get_bank_ids}
\title{Get bank IDs for target banks}
\usage{
get_bank_ids(
  bank_names,
  bank_meta_data,
  entity_type = "FHD",
  most_recent = TRUE
)
}
\arguments{
\item{bank_names}{The names of the banks to be selected from the \code{bank_meta_data}}

\item{bank_meta_data}{a \code{data.frame} with columns for ID_RSSID, Entity_Type, Name}

\item{entity_type}{The target entity type (Default: "FHD").  \code{NA} indicates the \code{entity_type} is not used.}

\item{most_recent}{if \code{TRUE} then select the most recent entry.}
}
\value{
If most_recent is \code{TRUE} then a vector of one ID for each bank_name.  Otherwise, multiple IDs may be returned in a list
}
\description{
Get bank IDs for target banks
}
\examples{
bank_meta_data <- data.frame(ID_RSSID = c(10, 20, 30),
  Entity_Type = rep("FHD", 3),
  Name = c("HUNTINGTON BANCSHARES INCORPORATED", "KEYCORP", "PNC BANK CORP"))
targets <- c("HUNTINGTON BANCSHARES INCORPORATED", "KEYCORP", "PNC BANK CORP")
target_ids <- get_bank_ids(targets, bank_meta_data)
all(target_ids == c(10, 20, 30))
}
