% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bank_meta_data.R
\name{get_bank_meta_data}
\alias{get_bank_meta_data}
\title{Get bank metadata from the Chicago Fed}
\usage{
get_bank_meta_data()
}
\value{
a \code{data.frame} containing the metadata
}
\description{
Get bank metadata from the Chicago Fed
}
\examples{
\dontrun{
get_bank_meta_data()
}
}
