% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fry9c_data.R
\name{get_fry9c_data}
\alias{get_fry9c_data}
\title{Get FR Y-9c data}
\usage{
get_fry9c_data(year, quarter, max_rows = 5000, verbose = TRUE)
}
\arguments{
\item{year}{the year requested starting in 1986}

\item{quarter}{the quarter requested [1,2,3,4]}

\item{max_rows}{the maximum number of rows to read from the file.  As of 2018, the files are generally less than 5,000 rows.  (Default: 5000)}

\item{verbose}{should messages be printed? (Default: TRUE)}
}
\value{
a \code{data.frame} containing the results
}
\description{
Query the Chicago Fed to obtain the requested year and quarter FR Y-9c
}
\examples{
\dontrun{X1Q2018 <- get_fry9c_data(2018, 1)}
}
