% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fry9c_historical_data.R
\name{get_fry9c_historical_data}
\alias{get_fry9c_historical_data}
\title{Get FR Y-9c data}
\usage{
get_fry9c_historical_data(year, quarter, verbose = TRUE)
}
\arguments{
\item{year}{the year requested starting in 1986}

\item{quarter}{the quarter requested [1,2,3,4]}

\item{verbose}{should messages be printed? (Default: TRUE)}
}
\value{
a \code{data.frame} containing the results
}
\description{
Query the Chicago Fed to obtain the requested year and quarter FR Y-9c  up to
2021 at \url{https://www.chicagofed.org/banking/financial-institution-reports/bhc-data}
}
\details{
After 2021, the data moved to \url{https://www.ffiec.gov/npw/FinancialReport/FinancialDataDownload}
}
\examples{
\dontrun{X <- get_fry9c_historical_data(2015, 1)}
}
