% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_fry9c.R
\name{parse_fry9c}
\alias{parse_fry9c}
\title{Parse a FR Y-9c XML schema}
\usage{
parse_fry9c(xml_filename, xsd_filename = NA)
}
\arguments{
\item{xml_filename}{The full filename of the FR Y-9c schema}

\item{xsd_filename}{(Default NA) An optional parameter defining the xml schema document}
}
\value{
an FRY9c object
}
\description{
Parse a FR Y-9c XML schema
}
\details{
Need to add ... to pass arguments to read_xml
}
\examples{
xml <- '
<FRY9C date="20160101" omb_number="1" title="Test FR Y-9C">
<schedule desig="A" title="B">
<component num="1." name="Name1" key="NA">
<component num="1.a." name="Name1a" key="NA">
</component>
</component>
<component num="2." name="Name2" key="NA">
</component>
</schedule>
</FRY9C>
'

tmpfilename <- tempfile(fileext = ".xml")
tmpfile <- file(tmpfilename, encoding = "UTF-8")
cat(xml, file = tmpfile)
close(tmpfile)
fry_9c <- parse_fry9c(tmpfilename)
unlink(tmpfilename)
}
