% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_fry9c.R
\name{parse_schedule}
\alias{parse_schedule}
\title{Parse the schedule portion of the FR Y-9c schema}
\usage{
parse_schedule(nodeset)
}
\arguments{
\item{nodeset}{an xml2 nodeset like that returned by xml_children}
}
\value{
an xml2 nodeset
}
\description{
Parse the schedule portion of the FR Y-9c schema
}
