% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\name{schedule}
\alias{schedule}
\alias{.schedule}
\alias{Schedule}
\title{Class providing an object to manipulate a schedule in a FR Y-9c}
\usage{
Schedule(desig, title)
}
\arguments{
\item{desig}{the designator of the schedule, e.g. "HI"}

\item{title}{the title of the schedule}
}
\value{
an object of class \code{schedule}
}
\description{
Class providing an object to manipulate a schedule in a FR Y-9c

Class providing an object to manipulate a schedule in a FR Y-9c
}
\examples{
x <- Schedule("HI", "Income Statement")
y <- Component("1.a.", "Income", "ZZZZ1234")
y$add(Component("1.a.(1)", "Sub-Income", "ABCD5555"))
x$add(y)
x$initializeData(data.frame(ZZZZ1234 = 1:4,
                            ABCD5555 = 5:8))
x$addBankNames(paste("bank", LETTERS[1:4], sep=""))
all(x$export_csv() == c("Schedule HI, Income Statement, ",
                        "1.a., Income, 1,2,3,4",
                        "1.a.(1), Sub-Income, 5,6,7,8"))
print(x)
all(x$getValueFromKey("ABCD5555") == 5:8)
all(x$getValueFromNum("1.a.") == 1:4)
x$getDesig() == "HI"
x$getTitle() == "Income Statement"
all(x$getBankNames() == paste("bank", LETTERS[1:4], sep=""))
x$getComponent(1)$getKey() == "ZZZZ1234"
all(x$sumLevels(c("1.a.", "1.a.(1)"), NA) == 1:4 + 5:8)
all(x$sumLevels("1.a.", "1.a.(1)") == 1:4 - 5:8)
all(x$createDataFrame()$bankA == c(1,5))
x$commonSize(100)
all.equal(x$getCommonSizeValueFromNum("1.a."), (1:4)/100)
all.equal(x$getCommonSizeValueFromKey("ZZZZ1234"), (1:4)/100)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{.schedule$new()}}
\item \href{#method-initializeData}{\code{.schedule$initializeData()}}
\item \href{#method-add}{\code{.schedule$add()}}
\item \href{#method-export_csv}{\code{.schedule$export_csv()}}
\item \href{#method-print}{\code{.schedule$print()}}
\item \href{#method-getMetricFrom}{\code{.schedule$getMetricFrom()}}
\item \href{#method-getComponentFrom}{\code{.schedule$getComponentFrom()}}
\item \href{#method-getValueFromKey}{\code{.schedule$getValueFromKey()}}
\item \href{#method-getNumFromKey}{\code{.schedule$getNumFromKey()}}
\item \href{#method-getKeyFromNum}{\code{.schedule$getKeyFromNum()}}
\item \href{#method-getValueFromNum}{\code{.schedule$getValueFromNum()}}
\item \href{#method-getCommonSizeValueFromNum}{\code{.schedule$getCommonSizeValueFromNum()}}
\item \href{#method-getCommonSizeValueFromKey}{\code{.schedule$getCommonSizeValueFromKey()}}
\item \href{#method-getComponentFromKey}{\code{.schedule$getComponentFromKey()}}
\item \href{#method-sumLevels}{\code{.schedule$sumLevels()}}
\item \href{#method-getDesig}{\code{.schedule$getDesig()}}
\item \href{#method-getTitle}{\code{.schedule$getTitle()}}
\item \href{#method-getBankNames}{\code{.schedule$getBankNames()}}
\item \href{#method-addBankNames}{\code{.schedule$addBankNames()}}
\item \href{#method-createDataFrame}{\code{.schedule$createDataFrame()}}
\item \href{#method-getComponent}{\code{.schedule$getComponent()}}
\item \href{#method-commonSize}{\code{.schedule$commonSize()}}
\item \href{#method-clone}{\code{.schedule$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize an object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.schedule$new(desig, title)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{desig}}{the designator of the schedule, e.g. "HI"}

\item{\code{title}}{the title of the schedule}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-initializeData"></a>}}
\if{latex}{\out{\hypertarget{method-initializeData}{}}}
\subsection{Method \code{initializeData()}}{
initialize the values in each \code{schedule}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.schedule$initializeData(dat)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dat}}{A dataset from the Fed with all FR Y-9c data for a quarter.  The dataset contains columns with names that correspond to \code{key}s}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add"></a>}}
\if{latex}{\out{\hypertarget{method-add}{}}}
\subsection{Method \code{add()}}{
add a \code{component} to this \code{schedule}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.schedule$add(comp)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{comp}}{a component}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-export_csv"></a>}}
\if{latex}{\out{\hypertarget{method-export_csv}{}}}
\subsection{Method \code{export_csv()}}{
export this \code{schedule} in CSV format
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.schedule$export_csv()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
print the \code{schedule} as a string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.schedule$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getMetricFrom"></a>}}
\if{latex}{\out{\hypertarget{method-getMetricFrom}{}}}
\subsection{Method \code{getMetricFrom()}}{
Get metric from key or num
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.schedule$getMetricFrom(key, num = NA, metric = .metric$value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{The lookup key associated with the line item number}

\item{\code{num}}{The line item number of a component of a schedule}

\item{\code{metric}}{the metric}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getComponentFrom"></a>}}
\if{latex}{\out{\hypertarget{method-getComponentFrom}{}}}
\subsection{Method \code{getComponentFrom()}}{
get component from key and num
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.schedule$getComponentFrom(key, num = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{The lookup key associated with the line item number}

\item{\code{num}}{The line item number of a component of a schedule}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getValueFromKey"></a>}}
\if{latex}{\out{\hypertarget{method-getValueFromKey}{}}}
\subsection{Method \code{getValueFromKey()}}{
get a \code{component} value from this object or a sub-\code{component} that matches the \code{key}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.schedule$getValueFromKey(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{The lookup key associated with the line item number}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getNumFromKey"></a>}}
\if{latex}{\out{\hypertarget{method-getNumFromKey}{}}}
\subsection{Method \code{getNumFromKey()}}{
get num from key
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.schedule$getNumFromKey(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{The lookup key associated with the line item number}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getKeyFromNum"></a>}}
\if{latex}{\out{\hypertarget{method-getKeyFromNum}{}}}
\subsection{Method \code{getKeyFromNum()}}{
get key from num
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.schedule$getKeyFromNum(num)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{num}}{The line item number of a component of a schedule}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getValueFromNum"></a>}}
\if{latex}{\out{\hypertarget{method-getValueFromNum}{}}}
\subsection{Method \code{getValueFromNum()}}{
get a \code{component} value from the \code{component} number in this object or a sub-\code{component}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.schedule$getValueFromNum(num)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{num}}{The line item number of a component of a schedule}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCommonSizeValueFromNum"></a>}}
\if{latex}{\out{\hypertarget{method-getCommonSizeValueFromNum}{}}}
\subsection{Method \code{getCommonSizeValueFromNum()}}{
get a \code{component} common-sized value from the \code{component} number in this object or a sub-\code{component}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.schedule$getCommonSizeValueFromNum(num)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{num}}{The line item number of a component of a schedule}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getCommonSizeValueFromKey"></a>}}
\if{latex}{\out{\hypertarget{method-getCommonSizeValueFromKey}{}}}
\subsection{Method \code{getCommonSizeValueFromKey()}}{
get a \code{component} common-sized value from the \code{component} \code{key} in this object or a sub-\code{component}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.schedule$getCommonSizeValueFromKey(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{The lookup key associated with the line item number}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getComponentFromKey"></a>}}
\if{latex}{\out{\hypertarget{method-getComponentFromKey}{}}}
\subsection{Method \code{getComponentFromKey()}}{
get component from key
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.schedule$getComponentFromKey(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{The lookup key associated with the line item number}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sumLevels"></a>}}
\if{latex}{\out{\hypertarget{method-sumLevels}{}}}
\subsection{Method \code{sumLevels()}}{
add or subtract the \code{component}s in this schedule
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.schedule$sumLevels(pos_num_list, neg_num_list)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pos_num_list}}{a vector of \code{component} \code{num}s to be added}

\item{\code{neg_num_list}}{a vector of \code{component} \code{num}s to be subtracted}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDesig"></a>}}
\if{latex}{\out{\hypertarget{method-getDesig}{}}}
\subsection{Method \code{getDesig()}}{
get the \code{desig} for this object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.schedule$getDesig()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTitle"></a>}}
\if{latex}{\out{\hypertarget{method-getTitle}{}}}
\subsection{Method \code{getTitle()}}{
get the \code{title} for this object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.schedule$getTitle()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getBankNames"></a>}}
\if{latex}{\out{\hypertarget{method-getBankNames}{}}}
\subsection{Method \code{getBankNames()}}{
get the \code{bank_names} assigned to this object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.schedule$getBankNames()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addBankNames"></a>}}
\if{latex}{\out{\hypertarget{method-addBankNames}{}}}
\subsection{Method \code{addBankNames()}}{
add \code{bank_names} to this object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.schedule$addBankNames(bank_names)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bank_names}}{a vector of bank names to be stored in the \code{schedule}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-createDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-createDataFrame}{}}}
\subsection{Method \code{createDataFrame()}}{
create a \code{data.frame} from the \code{component}s in this object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.schedule$createDataFrame()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getComponent"></a>}}
\if{latex}{\out{\hypertarget{method-getComponent}{}}}
\subsection{Method \code{getComponent()}}{
get a \code{component} by its index in the \code{schedule}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.schedule$getComponent(index)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{the index of the component}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-commonSize"></a>}}
\if{latex}{\out{\hypertarget{method-commonSize}{}}}
\subsection{Method \code{commonSize()}}{
common-size the \code{component}s in this \code{schedule} using the \code{divisor}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.schedule$commonSize(divisor)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{divisor}}{common size divisor}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.schedule$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
