% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fry9c.R
\name{fry9c}
\alias{fry9c}
\alias{.fry9c}
\alias{Fry9c}
\title{Class providing an object to manipulate a FR Y-9c and component schedules}
\usage{
Fry9c(date, omb_number, title)
}
\arguments{
\item{date}{The date associated with the FR Y-9c}

\item{omb_number}{The OMB number of the FR Y-9c documentation.  Something like "7100-0128"}

\item{title}{The title of the FR Y-9c.  Something like "Consolidated Financial Statements for Holding Companies--FR Y-9c"}
}
\value{
an object of class \code{fry9c}
}
\description{
Class providing an object to manipulate a FR Y-9c and component schedules

Class providing an object to manipulate a FR Y-9c and component schedules
}
\examples{
f <- Fry9c("20180331", "7100-0128",
           "Consolidated Financial Statements for Holding Companies--FR Y-9c")
x <- Schedule("HI", "Income Statement")
y <- Component("1.a.", "Income", "ZZZZ1234")
y$add(Component("1.a.(1)", "Sub-Income", "ABCD5555"))
x$add(y)
f$add(x)
f$initializeData(data.frame(ZZZZ1234 = 1:4,
                            ABCD5555 = 5:8))
print(f)
f$getSchedule("HI")$getDesig() == "HI"
f$addBankNames(paste("bank", LETTERS[1:4], sep=""))
\dontrun{
  f$exportExcel(tempfile(fileext = ".xlsx"))
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{.fry9c$new()}}
\item \href{#method-add}{\code{.fry9c$add()}}
\item \href{#method-getSchedule}{\code{.fry9c$getSchedule()}}
\item \href{#method-initializeData}{\code{.fry9c$initializeData()}}
\item \href{#method-print}{\code{.fry9c$print()}}
\item \href{#method-addBankNames}{\code{.fry9c$addBankNames()}}
\item \href{#method-exportExcel}{\code{.fry9c$exportExcel()}}
\item \href{#method-getDate}{\code{.fry9c$getDate()}}
\item \href{#method-getOmbNumber}{\code{.fry9c$getOmbNumber()}}
\item \href{#method-getTitle}{\code{.fry9c$getTitle()}}
\item \href{#method-clone}{\code{.fry9c$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
initialize the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.fry9c$new(date, omb_number, title)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{date}}{The date associated with the FR Y-9c}

\item{\code{omb_number}}{The OMB number of the FR Y-9c documentation.  Something like "7100-0128"}

\item{\code{title}}{The title of the FR Y-9c.  Something like "Consolidated Financial Statements for Holding Companies--FR Y-9c"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add"></a>}}
\if{latex}{\out{\hypertarget{method-add}{}}}
\subsection{Method \code{add()}}{
add a new \code{schedule}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.fry9c$add(sched)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sched}}{a schedule}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getSchedule"></a>}}
\if{latex}{\out{\hypertarget{method-getSchedule}{}}}
\subsection{Method \code{getSchedule()}}{
Get a \code{schedule} by the \code{desig}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.fry9c$getSchedule(desig)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{desig}}{An identifier for a component \code{schedule}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-initializeData"></a>}}
\if{latex}{\out{\hypertarget{method-initializeData}{}}}
\subsection{Method \code{initializeData()}}{
initialize the values in each \code{schedule} and \code{component}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.fry9c$initializeData(dat)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dat}}{A dataset from the Fed with all FR Y-9c data for a quarter.  The dataset contains columns with names that correspond to \code{key}s}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
print the \code{schedule}s and \code{component}s as a string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.fry9c$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addBankNames"></a>}}
\if{latex}{\out{\hypertarget{method-addBankNames}{}}}
\subsection{Method \code{addBankNames()}}{
add \code{bank_names} to this object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.fry9c$addBankNames(bank_names)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bank_names}}{a vector of bank names to be stored in the \code{schedule}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-exportExcel"></a>}}
\if{latex}{\out{\hypertarget{method-exportExcel}{}}}
\subsection{Method \code{exportExcel()}}{
export an Excel file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.fry9c$exportExcel(output_file_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{output_file_name}}{output file name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getDate"></a>}}
\if{latex}{\out{\hypertarget{method-getDate}{}}}
\subsection{Method \code{getDate()}}{
Get the Date
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.fry9c$getDate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getOmbNumber"></a>}}
\if{latex}{\out{\hypertarget{method-getOmbNumber}{}}}
\subsection{Method \code{getOmbNumber()}}{
get the OMB number
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.fry9c$getOmbNumber()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getTitle"></a>}}
\if{latex}{\out{\hypertarget{method-getTitle}{}}}
\subsection{Method \code{getTitle()}}{
Get the title
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.fry9c$getTitle()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.fry9c$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
