% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fry9c_group.R
\docType{class}
\name{fry9c_group}
\alias{fry9c_group}
\alias{.fry9c_group}
\alias{Fry9c_group}
\alias{length.fry9c_group}
\title{Class providing an object to manipulate a group of FR Y-9c templates, typically
used to manage FY Y-9c across years}
\format{\code{\link{R6Class}} object.}
\usage{
.fry9c_group

Fry9c_group(years, quarters)

\method{length}{fry9c_group}(x, ...)
}
\arguments{
\item{years}{The years associated with the fry9c objects in the group}

\item{quarters}{The quarters associate with the fry9c objects in the group}

\item{x}{the \code{fry9c_group} object}

\item{...}{not used}
}
\value{
Object of \code{\link{R6Class}}

an object of class \code{fry9c_group}

the number of \code{fry9c} objects in the group
}
\description{
Class providing an object to manipulate a group of FR Y-9c templates, typically
used to manage FY Y-9c across years
}
\section{Fields}{

\describe{
\item{\code{fry9c_list}}{a list of FR Y-9c objects}

\item{\code{years}}{The year associated with FR Y-9c in the list}

\item{\code{quarters}}{The quarter associated with each FR Y-9c in the list}
}}

\section{Methods}{

\describe{
  \item{Documentation}{}
  \item{\code{new(years, quarters)}}{Creates object of this class with room for FR Y-9cs associated with each \code{year} adn \code{quarter}}
  \item{\code{parse_fry9c(files)}}{Parse a collection of \code{files} which each represent a FR Y-9c schema.}
  \item{\code{initializeData(data_list, banks)}}{Parse a collection of FR Y-9c data files in \code{data_list}.  Also include a vector of \code{banks} that includes the names of the banks associated with the rows as they should appear in output.}
  \item{\code{get_fry9c(year, quarter)}}{Extract a \code{fry9c} object from the collection associated with a \code{year} and \code{quarter}}
  \item{\code{commonSize(divisor_sched, divisor_key, sched)}}{Common side the \code{sched} using the element from the \code{divisor_sched} and \code{divisor_key}.  This is usually done by common sizing an income statement with the average assets.}
  \item{\code{get_plot_data(sched, key)}}{Create a \code{data.frame} that can be used for plotting using \code{ggplot2} by selecting a schedule \code{sched} and element number \code{key}}
  \item{\code{length()}}{Return the number of fry9c objects in the collection}
  \item{\code{print()}}{Print a summary of the collection contents}
}
}

\examples{
# load example data
fry9c_data_list <- list(
  read.csv(system.file(file.path("extdata", "ex_BHCF1712.csv"), package = "fry9c")),
  read.csv(system.file(file.path("extdata", "ex_BHCF1812.csv"), package = "fry9c")))

my_fry9c_group <- Fry9c_group(years = c(2017, 2016),
                              quarters = c(4, 4))
my_fry9c_group$parse_fry9c(
  system.file(file.path("extdata", c("FR_Y-9C20171231.xml", "FR_Y-9C20161231.xml")),
              package = "fry9c"))

my_fry9c_group$initializeData(fry9c_data_list, paste("bank", LETTERS[1:10], sep=""))
print(my_fry9c_group)
length(my_fry9c_group) == 2

class(my_fry9c_group$get_fry9c(2016, 4))[1] == "fry9c"

my_fry9c_group$commonSize("HC-K", "BHCK3368", "HI") # 5.

nrow(my_fry9c_group$get_plot_data("HC-K", "BHCK3368")) == 20 # 5.
}
\keyword{datasets}
