% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\docType{class}
\name{schedule}
\alias{schedule}
\alias{Schedule}
\title{Class providing an object to manipulate a schedule in a FR Y-9c}
\format{\code{\link{R6Class}} object.}
\usage{
schedule

Schedule(desig, title)
}
\arguments{
\item{desig}{the designator of the schedule, e.g. "HI"}

\item{title}{the title of the schedule}
}
\value{
Object of \code{\link{R6Class}}

an object of class \code{schedule}
}
\description{
Class providing an object to manipulate a schedule in a FR Y-9c
}
\section{Fields}{

\describe{
\item{\code{desig}}{the designator of the schedule, e.g. "HI"}

\item{\code{title}}{the title of the schedule}
}}

\section{Methods}{

\describe{
  \item{Documentation}{}
  \item{\code{new(desig, title)}}{}
  \item{\code{initializeData(dat)}}{}
  \item{\code{add(comp)}}{}
  \item{\code{export_csv()}}{}
  \item{\code{print()}}{}
  \item{\code{getValueFromKey(key)}}{}
  \item{\code{getValueFromNum(num)}}{}
  \item{\code{getCommonSizeValueFromNum(num)}}{}
  \item{\code{sumLevels(pos_num_list, neg_num_list)}}{}
  \item{\code{getDesig()}}{}
  \item{\code{getTitle()}}{}
  \item{\code{getBankNames()}}{}
  \item{\code{addBankNames(bank_names)}}{}
  \item{\code{createDataFrame()}}{}
  \item{\code{getComponent(index)}}{}
  \item{\code{commonSize(divisor)}}{}
}
}

\examples{
x <- schedule$new("HI", "Income Statement")
Schedule("HI", "Income Statement")
}
\keyword{datasets}
