% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R
\docType{class}
\name{schedule}
\alias{schedule}
\alias{.schedule}
\alias{Schedule}
\title{Class providing an object to manipulate a schedule in a FR Y-9c}
\format{\code{\link{R6Class}} object.}
\usage{
.schedule

Schedule(desig, title)
}
\arguments{
\item{desig}{the designator of the schedule, e.g. "HI"}

\item{title}{the title of the schedule}
}
\value{
Object of \code{\link{R6Class}}

an object of class \code{schedule}
}
\description{
Class providing an object to manipulate a schedule in a FR Y-9c
}
\section{Fields}{

\describe{
\item{\code{desig}}{the designator of the schedule, e.g. "HI"}

\item{\code{title}}{the title of the schedule}

\item{\code{dat}}{A dataset from the Fed with all FR Y-9c data for a quarter.  The dataset contains columns with names that correspond to \code{key}s}

\item{\code{comp}}{a \code{component}}

\item{\code{num}}{The line item number of a component of a schedule}

\item{\code{key}}{The lookup key associated with the line item number}

\item{\code{index}}{the index of the desired \code{component} in the \code{schedule}}

\item{\code{bank_names}}{a vector of bank names to be stored in the \code{schedule}}

\item{\code{pos_num_list}}{a vector of \code{component} \code{num}s to be added}

\item{\code{neg_num_list}}{a vector of \code{component} \code{num}s to be subtracted}
}}

\section{Methods}{

\describe{
  \item{Documentation}{}
  \item{\code{new(desig, title)}}{generate a new \code{schedule}}
  \item{\code{initializeData(dat)}}{initialize the values in each \code{schedule}}
  \item{\code{add(comp)}}{add a \code{component} to this \code{schedule}}
  \item{\code{export_csv()}}{export this \code{schedule} in CSV format}
  \item{\code{print()}}{print the \code{schedule} as a string}
  \item{\code{getValueFromKey(key)}}{get a \code{component} value from this object or a sub-\code{component} that matches the \code{key}}
  \item{\code{getValueFromNum(num)}}{get a \code{component} value from the \code{component} number in this object or a sub-\code{component}}
  \item{\code{getCommonSizeValueFromNum(num)}}{get a \code{component} common-sized value from the \code{component} number in this object or a sub-\code{component}}
  \item{\code{sumLevels(pos_num_list, neg_num_list)}}{add or subtract the \code{component}s in this schedule}
  \item{\code{getDesig()}}{get the \code{desig} for this object}
  \item{\code{getTitle()}}{get the \code{title} for this object}
  \item{\code{getBankNames()}}{get the \code{bank_names} assigned to this object}
  \item{\code{addBankNames(bank_names)}}{add \code{bank_names} to this object}
  \item{\code{createDataFrame()}}{create a \code{data.frame} from the \code{component}s in this object}
  \item{\code{getComponent(index)}}{get a \code{component} by its index in the \code{schedule}}
  \item{\code{commonSize(divisor)}}{common-size the \code{component}s in this \code{schedule} using the \code{divisor}}
}
}

\examples{
x <- Schedule("HI", "Income Statement")
y <- Component("1.a.", "Income", "ZZZZ1234")
y$add(Component("1.a.(1)", "Sub-Income", "ABCD5555"))
x$add(y)
x$initializeData(data.frame(ZZZZ1234 = 1:4,
                            ABCD5555 = 5:8))
x$addBankNames(paste("bank", LETTERS[1:4], sep=""))
all(x$export_csv() == c("Schedule HI, Income Statement, ",
                        "1.a., Income, 1,2,3,4",
                        "1.a.(1), Sub-Income, 5,6,7,8"))
print(x)
all(x$getValueFromKey("ABCD5555") == 5:8)
all(x$getValueFromNum("1.a.") == 1:4)
x$getDesig() == "HI"
x$getTitle() == "Income Statement"
all(x$getBankNames() == paste("bank", LETTERS[1:4], sep=""))
x$getComponent(1)$getKey() == "ZZZZ1234"
all(x$sumLevels(c("1.a.", "1.a.(1)"), NA) == 1:4 + 5:8)
all(x$sumLevels("1.a.", "1.a.(1)") == 1:4 - 5:8)
all(x$createDataFrame()$bankA == c(1,5))
x$commonSize(100)
all.equal(x$getCommonSizeValueFromNum("1.a."), (1:4)/100)
}
\keyword{datasets}
